fix: Use atomic operations for evalDepth to prevent race condition
The evalDepth counter was being incremented/decremented without any
synchronization, causing a data race when async expressions spawn
goroutines that call back into EvaluateExpression concurrently.

Replace direct int access with atomic.AddInt64 operations and change
the field type from int to int64.

Closes #119

+++ pkg/interpreter/interpreter.go
@@ -37,7 +37,7 @@ type Interpreter struct {
 	constants        map[string]struct{}      // Tracks names that are constants (immutable)
 	contracts        map[string]ContractDef   // Contract definitions by name
 	traitDefs        map[string]TraitDef      // Trait definitions by name
-	evalDepth        int                      // Current recursion depth for evaluation
+	evalDepth        int64                    // Current recursion depth for evaluation (atomic)
 }
 
 // NewInterpreter creates a new interpreter instance
